requirements.txt
asgiref==3.7.2
attrs==23.1.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.6.2
Brlapi==0.8.4
certifi==2022.9.24
chardet==5.1.0
click==8.1.6
cloud-init==23.3.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.8.0
distro-info==1.5
dj-database-url==2.1.0
-e git+https://github.com/django/django.git@b0ec87b8578147be4357c90eabcd2b916c780810#egg=Django
django-debug-toolbar==4.2.0
duplicity==1.2.2
fasteners==0.17.3
freetype-py==2.4.0
future==0.18.2
gunicorn==21.2.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.2.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.26.0
Mako==1.2.4.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
package-name==0.1
packaging==23.2
paramiko==2.12.0
pexpect==4.8.0
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.24.0
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.46.0
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.6.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.0.4
requests==2.31.0
rich==13.3.1
rlPyCairo==0.3.0
SecretStorage==3.3.3
six==1.16.0
soundconverter==4.0.4
sqlparse==0.4.4
systemd-python==235
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.2
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.16
wadllib==1.3.6
whitenoise==6.6.0
xdg==5
xkit==0.0.0
zipp==1.0.0
mysql-connector-python==8.0.26